[oilshell] Improve error message and remove unused code.
yajl.dump(obj, f) was not reliable.  Somehow it didn't raise errors
seemed to corrupt the Python interpreter.

It did not do any streaming either!  It just made a big string and
called write(), which we can do in the application.

So remove that, and then Oil always uses yajl.dumps(obj).  We need to
switch to a different binding later.